Fix hecke operator precision
The fix computes T(n)f with lower precision prec/Norm(n). To make sure this is enough precision to uniquely determine the modular form, we reconstruct it with a basis for the space. Since loading a basis can take time for high weight, there is also a version of the function that takes in a precomputed basis.