KodairaDimension
Fixed a sign error, so that CoefficientComponent runs correctly.

Let F(t) be a rational function representing a Hilbert modular form (or a quotient thereof). Write F(t) = sum_{n >> n0} a(n) t^n. If a(12n) = p(n) for some polynomial p and all n sufficiently large, KodairaDimension(F) returns the degree of p. I believe that in particular contexts, this is precisely the Kodaira dimension of the Hilbert modular variety. But this function may need to be renamed and incorporated into larger aspects of the code.